[svn] Allow decimal values for --timeout, --wait, and --waitretry. Message-ID: <m3llsjhx6t.fsf@hniksic.iskon.hr>

+++ src/connect.h
@@ -39,7 +39,7 @@ int connect_to_many PARAMS ((struct address_list *, unsigned short, int));
 void set_connection_host_name PARAMS ((const char *));
 
 int test_socket_open PARAMS ((int));
-int select_fd PARAMS ((int, int, int));
+int select_fd PARAMS ((int, double, int));
 uerr_t bindport PARAMS ((unsigned short *, int));
 uerr_t acceptport PARAMS ((int *));
 void closeport PARAMS ((int));
